% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_columns.R
\name{convert_date_columns}
\alias{convert_date_columns}
\title{Covert a date column to separate values}
\usage{
convert_date_columns(data, date_column)
}
\arguments{
\item{data}{data file}

\item{date_column}{specific column in the data which you want to convert to 3 separate columns}
}
\value{
your data with 3 separate columns for year, month and date instead of one column
}
\description{
This function can be used to convert a date column with the yy/mm/dd
format to separate columns named year, month and date.
}
\examples{
# modified_data <- convert_date_columns(data, "date")
}
