% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\title{Plotting data with a group variable}
\usage{
plotting(data, x_var, y_var, group_var, title)
}
\arguments{
\item{data}{data file}

\item{x_var}{variable you want to have on the x-axis}

\item{y_var}{variable you want to have on the y-axis}

\item{group_var}{the group variable}

\item{title}{title of the plot}
}
\value{
a plot of your data with x-axis, y-axis, group variable and title
}
\description{
This function can be used to make a plot of occurrences with a group variable
}
\examples{
# plotting(data = joined_data, year, searches_flu, continent, "Flu occurrences for each continent")

}
