% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_tibble.R
\name{read_csv_tibble}
\alias{read_csv_tibble}
\title{Read csv file}
\usage{
read_csv_tibble(file_path)
}
\arguments{
\item{file_path}{path to the csv file you want to use}
}
\value{
the data from your csv file in a tibble format
}
\description{
This function can be used to read a csv file from your folder and convert
it into a tibble
}
\examples{
# data <- read_csv_tibble("path/to/your/file.csv")
}
