% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds_csv.R
\name{save_rds_csv}
\alias{save_rds_csv}
\title{Save data as an .rds and .csv file}
\usage{
save_rds_csv(data, file_name)
}
\arguments{
\item{data}{data file}

\item{file_name}{name of the output file}
}
\value{
your data in .rds and .csv files
}
\description{
This function can be used to save your data as an .rds or .csv file
}
\examples{
# data <- read_excel_data("path/to/data.xlsx", "Sheet1")
}
